+++ classes/Box.h
@@ -8,29 +8,29 @@ class Box
 {
 public:
 	Box(void);
-	Box(int width, int height, int depth);
-	Box(int width,int height, int depth, std::vector<Cell>);
+	Box(float width, float height, float depth);
+	Box(float width,float height, float depth, std::vector<Cell>);
 	~Box(void);
-	int getWidth() const;
-	int getHeight() const;
-	int getDepth() const;
+	float getWidth() const;
+	float getHeight() const;
+	float getDepth() const;
 	std::vector<Cell*> getCells() const;
-	void setWidth(int width);
-	void setHeight(int height);
-	void setDepth(int depth);
+	void setWidth(float width);
+	void setHeight(float height);
+	void setDepth(float depth);
 	void setCells(std::vector<Cell*> cells);
 	void addCell(Cell* cell);
 	void printBox() const;
 	void deleteCells();
-	void setSize(int w,int h,int d);
-	void reduceISO(const int pas);
-	void reduceWidth(const int pas);
-	void reduceHeight(const int pas);
-	void reduceDepth(const int pas);
+	void setSize(float w,float h,float d);
+	void reduceISO(const float pas);
+	void reduceWidth(const float pas);
+	void reduceHeight(const float pas);
+	void reduceDepth(const float pas);
 	void updateForces();
 	void applyForces();
 private:
-	int width, height, depth;
+	float width, height, depth;
 	std::vector<Cell*> cells;
 };
